refactor logics and validations

+++ src/interfaces/IController.sol
@@ -9,7 +9,7 @@ interface IController {
         external
         returns (DataType.TradeResult memory);
 
-    function updateMargin(uint256 _pairGroupId, int256 _marginAmount) external returns (uint256 vaultId);
+    function updateMargin(uint64 _pairGroupId, int256 _marginAmount) external returns (uint256 vaultId);
 
     function getSqrtPrice(uint256 _pairId) external view returns (uint160);
